+++ src/plugins/i18n.ts
@@ -1,0 +1,3 @@
+export {
+  createI18nPlugin,
+} from "./i18n/plugin.js";